PR-E-3: extract WellKnownReferences (BCL ref fields + lazy initializers)
Third step in the ReflectionMetadataEmitter decomposition. Moves the ~25
well-known BCL TypeReferenceHandle / MemberReferenceHandle fields and
their paired Get* lazy initializer methods out of ReflectionMetadataEmitter
into a new internal sealed class WellKnownReferences.

Fields moved (now state on WellKnownReferences):
  Eager (set in WellKnownReferences's constructor):
  - ObjectTypeRef
  - ValueTypeRef
  - ObjectCtorRef
  Lazy MemberReferenceHandle? (cached on first Get*):
  - notImplementedExceptionCtorRef
  - delegateCombineRef, delegateRemoveRef
  - interlockedCompareExchangeOpenRef
  - isReadOnlyAttributeCtorRef, isByRefLikeAttributeCtorRef
  - obsoleteAttributeStringBoolCtorRef
  - systemAttributeTypeRef, systemAttributeCtorRef
  Lazy MemberReferenceHandle (IsNil-checked):
  - stringConcatRef, stringEqualsRef
  - objectStaticEqualsRef
  - objectInstanceToStringRef, objectInstanceGetHashCodeRef
  - nullRefExceptionCtorRef
  - stringConcatArrayRef
  - convertToStringRef, cultureInvariantGetterRef
  - hashCodeAddOpenRef, hashCodeToHashCodeRef
  - hashCodeTypeRef (TypeReferenceHandle)
  Array-shaped cache:
  - hashCodeCombineOpenRefs[8]

Methods moved (now public on WellKnownReferences, same shape):
  GetNotImplementedExceptionCtor
  GetDelegateCombineRef, GetDelegateRemoveRef
  GetInterlockedCompareExchangeOpenRef
  GetIsReadOnlyAttributeTypeRef, GetIsReadOnlyAttributeCtorRef
  GetIsByRefLikeAttributeCtorRef
  GetObsoleteAttributeStringBoolCtorRef
  GetSystemAttributeTypeRef, GetSystemAttributeCtorRef
  GetNullReferenceExceptionCtorRef
  GetNullableGetValueReference, GetNullableGetHasValueReference
  GetStringLengthReference, GetStringCharsReference
  GetTypeFromHandleReference, GetArrayCopyReference
  GetStringConcatReference, GetStringEqualsReference
  GetObjectInstanceToStringReference, GetObjectInstanceGetHashCodeReference
  GetObjectStaticEqualsReference
  GetHashCodeTypeReference
  GetHashCodeCombineOpenReference, GetHashCodeAddOpenReference
  GetHashCodeToHashCodeReference
  GetConvertToStringReference, GetCultureInvariantGetterReference
  GetStringConcatArrayReference

The eager-init block that lived inline in ReflectionMetadataEmitter.EmitCore
(directly calling metadata.AddTypeReference / AddMemberReference for
ObjectTypeRef / ValueTypeRef / ObjectCtorRef) is now relocated into
WellKnownReferences's constructor.

Design:
- WellKnownReferences depends on EmitContext (Metadata, References, Core*)
  and on two delegate hooks (Func<Type, TypeReferenceHandle> and
  Func<MethodInfo, MemberReferenceHandle>) that route to the
  ReflectionMetadataEmitter dedup-cached resolvers (which themselves use
  MetadataTokenCache). This satisfies the constraint that
  WellKnownReferences MUST NOT depend on ReflectionMetadataEmitter itself.
- Eager refs exposed as get-only properties; lazy refs exposed as Get*()
  methods with identical shape to the originals.
- GetObjectDefaultCtorReference()'s factory body is now a private
  BuildObjectDefaultCtorReference() used once to populate ObjectCtorRef
  at construction time; consumers read wellKnown.ObjectCtorRef.
- hashCodeLocalSig + GetHashCodeLocalSignature stay on the emitter (it
  returns StandaloneSignatureHandle and belongs to later DataStructSynthesizer).
- GetHashCodeCombineObjectSpec / GetHashCodeAddObjectSpec stay on the
  emitter (they construct MethodSpecs); both now call into
  wellKnown.GetHashCodeCombineOpenReference / AddOpenReference.

Verification:
- src/Core/CodeAnalysis/Emit/WellKnownReferences.cs exists (827 LoC).
- ReflectionMetadataEmitter.cs LoC: 17873 -> 17217 (-656).
- RefactoringBaselineTests: 92/92 pass with zero hash diffs.
- DeterministicEmitTests: 3/3 pass.
- Full Core.Tests suite: 2120/2120 pass.
- No public/internal API visible to Compilation.cs or any test was renamed.

Next PR in the sequence: PR-E-4 SlotPlanner.
